// material-kit-react/src/store/types.ts
export enum DataDictionaryActionType {
    GET_DATA_DICTIONARY = 'GET_DATA_DICTIONARY',
    GET_DATA_DICTIONARY_SUCCESS = 'GET_DATA_DICTIONARY_SUCCESS',
    GET_DATA_DICTIONARY_ERROR = 'GET_DATA_DICTIONARY_ERROR',

    CREATE_DATASET = 'CREATE_DATASET',
    CREATE_DATASET_SUCCESS = 'CREATE_DATASET_SUCCESS',
    CREATE_DATASET_ERROR = 'CREATE_DATASET_ERROR',



    GET_CREATE_DATASET_STATUS = 'GET_CREATE_DATASET_STATUS',
    GET_CREATE_DATASET_STATUS_SUCCESS = 'GET_CREATE_DATASET_STATUS_SUCCESS',
    GET_CREATE_DATASET_STATUS_ERROR = 'GET_CREATE_DATASET_STATUS_ERROR',


    UPLOAD_DATASET_FILES = 'UPLOAD_DATASET_FILES',
    UPLOAD_DATASET_FILES_SUCCESS = 'UPLOAD_DATASET_FILES_SUCCESS',
    UPLOAD_DATASET_FILES_ERROR = 'UPLOAD_DATASET_FILES_ERROR',



    DELETE_DATASET = 'DELETE_DATASET',
    DELETE_DATASET_SUCCESS = 'DELETE_DATASET_SUCCESS',
    DELETE_DATASET_ERROR = 'DELETE_DATASET_ERROR',

    
    GET_DATASET_FOLDER_LIST = 'GET_DATASET_FOLDER_LIST',
    GET_DATASET_FOLDER_LIST_SUCCESS = 'GET_DATASET_FOLDER_LIST_SUCCESS',
    GET_DATASET_FOLDER_LIST_ERROR = 'GET_DATASET_FOLDER_LIST_ERROR',

    UPDATE_DATASET_FOLDER_LIST = 'UPDATE_DATASET_FOLDER_LIST',
    UPDATE_DATASET_FOLDER_LIST_SUCCESS = 'UPDATE_DATASET_FOLDER_LIST_SUCCESS',
    UPDATE_DATASET_FOLDER_LIST_ERROR = 'UPDATE_DATASET_FOLDER_LIST_ERROR',

    DATASET_FOLDER_FILE_TREE_TOGGLE = 'DATASET_FOLDER_FILE_TREE_TOGGLE',

    DATASET_SELECT_FOLDER_TOGGLE = 'DATASET_SELECT_FOLDER_TOGGLE',
    DATASET_SELECT_FOLDER_TOGGLE_SUCCESS = 'DATASET_SELECT_FOLDER_TOGGLE_SUCCESS',
    DATASET_SELECT_FOLDER_TOGGLE_ERROR = 'DATASET_SELECT_FOLDER_TOGGLE_ERROR',

    DATASET_SELECT_FOLDER_CLEAR = 'DATASET_SELECT_FOLDER_CLEAR',


    DATASET_SELECT_FILE_VIEW = 'DATASET_SELECT_FILE_VIEW',

    DELETE_DATASET_FOLDER_LIST = 'DELETE_DATASET_FOLDER_LIST',
    DELETE_DATASET_FOLDER_LIST_SUCCESS = 'DELETE_DATASET_FOLDER_LIST_SUCCESS',
    DELETE_DATASET_FOLDER_LIST_ERROR = 'DELETE_DATASET_FOLDER_LIST_ERROR',

    GET_SELECT_FOLER_PATH = 'GET_SELECT_FOLER_PATH',
    GET_SELECT_FOLER_PATH_SUCCESS = 'GET_SELECT_FOLER_PATH_SUCCESS',
    GET_SELECT_FOLER_PATHT_ERROR = 'GET_SELECT_FOLER_PATHT_ERROR',

    ADD_IDS_FILES_TO_DATASET = 'ADD_IDS_FILES_TO_DATASET',
    ADD_IDS_FILES_TO_DATASET_SUCCESS = 'ADD_IDS_FILES_TO_DATASET_SUCCESS',
    ADD_IDS_FILES_TO_DATASET_ERROR = 'ADD_IDS_FILES_TO_DATASET_ERROR',

    SUBMIT_DATASET_FILES = 'SUBMIT_DATASET_FILES',
    SUBMIT_DATASET_FILES_SUCCESS = 'SUBMIT_DATASET_FILES_SUCCESS',
    SUBMIT_DATASET_FILES_ERROR = 'SUBMIT_DATASET_FILES_ERROR',

    CLEAR_DATASET_FILES_CACHE = 'CLEAR_DATASET_FILES_CACHE',



}